ECS strap component
