+++ Content.Client/Buckle/BuckleComponent.cs
@@ -1,11 +1,11 @@
 using Content.Shared.Buckle.Components;
 
-namespace Content.Client.Buckle
+namespace Content.Client.Buckle;
+
+[RegisterComponent]
+[ComponentReference(typeof(SharedBuckleComponent))]
+[Access(typeof(BuckleSystem))]
+public sealed class BuckleComponent : SharedBuckleComponent
 {
-    [RegisterComponent]
-    [ComponentReference(typeof(SharedBuckleComponent))]
-    public sealed class BuckleComponent : SharedBuckleComponent
-    {
-        public int? OriginalDrawDepth { get; set; }
-    }
+    public int? OriginalDrawDepth { get; set; }
 }
