+++ Content.Client/Buckle/Strap/StrapComponent.cs
@@ -1,15 +1,10 @@
 ﻿using Content.Shared.Buckle.Components;
-using Content.Shared.DragDrop;
 
-namespace Content.Client.Buckle.Strap
+namespace Content.Client.Buckle.Strap;
+
+[RegisterComponent]
+[ComponentReference(typeof(SharedStrapComponent))]
+[Access(typeof(BuckleSystem))]
+public sealed class StrapComponent : SharedStrapComponent
 {
-    [RegisterComponent]
-    [ComponentReference(typeof(SharedStrapComponent))]
-    public sealed class StrapComponent : SharedStrapComponent
-    {
-        public override bool DragDropOn(DragDropEvent eventArgs)
-        {
-            return false;
-        }
-    }
 }
